# feat(query): inverted index use empty position data when query not contain phrase terms