… Cargo.lock
… Cargo.toml
… src/common/exception/Cargo.toml
… src/query/ee/Cargo.toml
… src/query/ee/tests/it/inverted_index/pruning.rs
… src/query/service/tests/it/storages/testdata/caches_table.txt
… src/query/storages/common/index/Cargo.toml
… src/query/storages/common/index/src/inverted_index.rs
… src/query/storages/common/index/src/lib.rs
… src/query/storages/fuse/Cargo.toml
… src/query/storages/fuse/src/io/read/meta/meta_readers.rs
… src/query/storages/fuse/src/io/write/inverted_index_writer.rs
… src/query/storages/fuse/src/pruning/block_pruner.rs
… src/query/storages/fuse/src/pruning/inverted_index_pruner.rs
… src/query/storages/system/src/caches_table.rs